More rider goodness

+++ src/CycleTracker.Data/Models/Bike.cs
@@ -5,7 +5,7 @@ namespace CycleTracker.Data.Models
 {
 	public class Bike : CycleTrackerBase
 	{
-		public int RiderId { get; set; }
+		public long RiderId { get; set; }
 		public string Make { get; set; }
 		public string Model { get; set; }
 		public int Year { get; set; }
